fix(ResultView): int 형 범위 넘었을 경우, 잘못된 계산 결과를 반환하는 오류 수정

+++ src/main/java/lotto/view/ResultView.java
@@ -69,8 +69,9 @@ public class ResultView {
     }
 
     private static double calculateProfitRate(List<Rank> matchResults, Money money) {
-        int totalWinningPrize = matchResults.stream()
-                .mapToInt(Rank::getWinningPrize).sum();
+        long totalWinningPrize = matchResults.stream()
+                .mapToLong(Rank::getWinningPrize)
+                .sum();
         return (totalWinningPrize * 1.0) / (money.getMoney() * 1.0);
     }
 }
